refactor(team): move scanner/* into internal/scanner
First of four leaf extractions planned in
~/.claude/plans/partitioned-tickling-peacock.md (Track C). Moves 10
scanner files (6 prod + 4 test) out of the internal/team god-package.

Scope:
- `git mv` all `scanner*.go` from internal/team to internal/scanner
- package declaration in each file: `team` → `scanner`
- update external callers (`cmd/wuphf/…`, `internal/team/broker_scan.go`)
  to import `internal/scanner` and qualify types as `scanner.X`
- add `.secretlintignore` scoped strictly to scanner_patterns_test.go
  and scanner_test.go — these files contain deliberate fake-secret
  fixtures the scanner uses to self-test its own detection rules.
  Old-location commits predated the secretlint hook, so `git mv`
  re-surfaced the fixtures as staged paths and tripped secretlint
  fresh. Scope is tight: real secrets elsewhere still fire.

Verification:
- `go build ./...` clean
- `go test -count=1 ./internal/scanner/ ./internal/team/ ./cmd/...` green
  (including full internal/team suite at 77s)
