+++ .secretlintignore
@@ -1,0 +1,9 @@
+# Scanner self-test fixtures contain deliberately-synthetic secret-shaped
+# strings that exercise the scanner's own detection rules — they are NOT
+# real credentials. If secretlint also catches them, the hook flags on
+# every unrelated commit that happens to touch these files.
+#
+# Keep the scope tight: only the scanner's pattern + end-to-end test
+# files. Real secrets accidentally committed elsewhere still fire.
+internal/scanner/scanner_patterns_test.go
+internal/scanner/scanner_test.go
